requirements.txt
PyYAML==6.0
tenacity==8.1.0
betfairlightweight==2.17.0
pymongo~=4.3.3
pandas~=1.5.3
python-dotenv~=0.21.1
streamlit==1.17.0
watchdog==2.2.1
numpy~=1.23.3
matplotlib~=3.6.2
dacite~=1.8.0
tomli==2.0.1
tomlkit==0.11.6
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocs-gen-files==0.4.0
mkdocs-literate-nav==0.6.0
mkdocs-material==9.0.2
mkdocs-material-extensions==1.1.1
mkdocs-section-index==0.3.4
mkdocstrings==0.19.1
mkdocstrings-python==0.8.3
isort==5.11.4
black==22.12.0
pre-commit==2.21.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.15.9